Renamed mirte_control
