+++ mirte_control/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.16)
-project(ros2_control_demo_example_2 LANGUAGES CXX)
+project(mirte_control LANGUAGES CXX)
 
 if(CMAKE_CXX_COMPILER_ID MATCHES "(GNU|Clang)")
   add_compile_options(-Wall -Wextra)
@@ -23,42 +23,42 @@ endforeach()
 
 ## COMPILE
 add_library(
-  ros2_control_demo_example_2
+  mirte_control
   SHARED
   hardware/diffbot_system.cpp
 )
-target_compile_features(ros2_control_demo_example_2 PUBLIC cxx_std_17)
-target_include_directories(ros2_control_demo_example_2 PUBLIC
+target_compile_features(mirte_control PUBLIC cxx_std_17)
+target_include_directories(mirte_control PUBLIC
 $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/hardware/include>
-$<INSTALL_INTERFACE:include/ros2_control_demo_example_2>
+$<INSTALL_INTERFACE:include/mirte_control>
 )
 ament_target_dependencies(
-  ros2_control_demo_example_2 PUBLIC
+  mirte_control PUBLIC
   ${THIS_PACKAGE_INCLUDE_DEPENDS}
 )
 
 # Causes the visibility macros to use dllexport rather than dllimport,
 # which is appropriate when building the dll but not consuming it.
-target_compile_definitions(${PROJECT_NAME} PRIVATE "ROS2_CONTROL_DEMO_EXAMPLE_2_BUILDING_DLL")
+target_compile_definitions(${PROJECT_NAME} PRIVATE "MIRTE_CONTROL_BUILDING_DLL")
 
 # Export hardware plugins
-pluginlib_export_plugin_description_file(hardware_interface ros2_control_demo_example_2.xml)
+pluginlib_export_plugin_description_file(hardware_interface mirte_control.xml)
 
 # INSTALL
 install(
   DIRECTORY hardware/include/
-  DESTINATION include/ros2_control_demo_example_2
+  DESTINATION include/mirte_control
 )
 install(
   DIRECTORY description/launch description/ros2_control description/urdf description/rviz
-  DESTINATION share/ros2_control_demo_example_2
+  DESTINATION share/mirte_control
 )
 install(
   DIRECTORY bringup/launch bringup/config
-  DESTINATION share/ros2_control_demo_example_2
+  DESTINATION share/mirte_control
 )
-install(TARGETS ros2_control_demo_example_2
-  EXPORT export_ros2_control_demo_example_2
+install(TARGETS mirte_control
+  EXPORT export_mirte_control
   ARCHIVE DESTINATION lib
   LIBRARY DESTINATION lib
   RUNTIME DESTINATION bin
@@ -69,6 +69,6 @@ if(BUILD_TESTING)
 endif()
 
 ## EXPORTS
-ament_export_targets(export_ros2_control_demo_example_2 HAS_LIBRARY_TARGET)
+ament_export_targets(export_mirte_control HAS_LIBRARY_TARGET)
 ament_export_dependencies(${THIS_PACKAGE_INCLUDE_DEPENDS})
 ament_package()
