+++ mirte_control/description/ros2_control/diffbot.ros2_control.xacro
@@ -5,7 +5,7 @@
 
     <ros2_control name="${name}" type="system">
       <hardware>
-        <plugin>ros2_control_demo_example_2/DiffBotSystemHardware</plugin>
+        <plugin>mirte_control/DiffBotSystemHardware</plugin>
         <param name="example_param_hw_start_duration_sec">0</param>
         <param name="example_param_hw_stop_duration_sec">0.1</param>
       </hardware>
