+++ mirte_control/description/urdf/diffbot.materials.xacro
