+++ mirte_control/description/urdf/diffbot.urdf.xacro
@@ -3,13 +3,13 @@
 <robot xmlns:xacro="http://www.ros.org/wiki/xacro" name="diffdrive_robot">
   <xacro:arg name="prefix" default="" />
 
-  <xacro:include filename="$(find ros2_control_demo_example_2)/urdf/diffbot_description.urdf.xacro" />
+  <xacro:include filename="$(find mirte_control)/urdf/diffbot_description.urdf.xacro" />
 
   <!-- Import Rviz colors -->
-  <xacro:include filename="$(find ros2_control_demo_example_2)/urdf/diffbot.materials.xacro" />
+  <xacro:include filename="$(find mirte_control)/urdf/diffbot.materials.xacro" />
 
   <!-- Import diffbot ros2_control description -->
-  <xacro:include filename="$(find ros2_control_demo_example_2)/ros2_control/diffbot.ros2_control.xacro" />
+  <xacro:include filename="$(find mirte_control)/ros2_control/diffbot.ros2_control.xacro" />
 
   <xacro:diffbot prefix="$(arg prefix)" />
