+++ mirte_control/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>ros2_control_demo_example_2</name>
+  <name>mirte_control</name>
   <version>0.0.0</version>
   <description>Demo package of `ros2_control` hardware for DiffBot.</description>
